hw3/requirements.txt
telethon==1.42.0
kafka-python==2.0.2
